// front/src/services/generated/models/ImageEditRegion.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Normalized rectangle on the original image; output outside it is locally preserved.
 */
export type ImageEditRegion = {
    'x': number;
    'y': number;
    width: number;
    height: number;
};